sync origin repositsy

+++ includes/learnopengl/mesh.h
@@ -73,11 +73,11 @@ public:
             if(name == "texture_diffuse")
                 number = std::to_string(diffuseNr++);
             else if(name == "texture_specular")
-                number = std::to_string(specularNr++); // transfer unsigned int to stream
+                number = std::to_string(specularNr++); // transfer unsigned int to string
             else if(name == "texture_normal")
-                number = std::to_string(normalNr++); // transfer unsigned int to stream
+                number = std::to_string(normalNr++); // transfer unsigned int to string
              else if(name == "texture_height")
-                number = std::to_string(heightNr++); // transfer unsigned int to stream
+                number = std::to_string(heightNr++); // transfer unsigned int to string
 
             // now set the sampler to the correct texture unit
             glUniform1i(glGetUniformLocation(shader.ID, (name + number).c_str()), i);
@@ -87,7 +87,7 @@ public:
         
         // draw mesh
         glBindVertexArray(VAO);
-        glDrawElements(GL_TRIANGLES, indices.size(), GL_UNSIGNED_INT, 0);
+        glDrawElements(GL_TRIANGLES, static_cast<unsigned int>(indices.size()), GL_UNSIGNED_INT, 0);
         glBindVertexArray(0);
 
         // always good practice to set everything back to defaults once configured.
